# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Discord Music Bot API
  version: 1.0.0
  description: |
    RESTful API for the Discord Music Bot system.

    Provides endpoints for:
    - Analytics and metrics
    - Guild management
    - Music queue operations
    - Premium subscriptions
    - Webhook integrations (Stripe, Discord)

    ## Authentication
    All endpoints require API key authentication via the `X-API-Key` header,
    except for webhook endpoints which use signature verification.

    ## Rate Limiting
    - General endpoints: 100 requests/minute
    - Search endpoints: 10 requests/minute
    - Webhook endpoints: 50 requests/minute

  contact:
    name: Discord Bot Team
    url: https://github.com/your-org/discord-bot
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.discordmusicbot.com/api/v1
    description: Production server (global)
  - url: https://staging-api.discordmusicbot.com/api/v1
    description: Pre-production / staging server
  - url: https://sandbox-api.discordmusicbot.com/api/v1
    description: Internal development sandbox (VPN only)
...
security:
  - ApiKeyAuth: []
paths:
...
  /analytics/guilds:
    get:
      tags:
        - Analytics
      summary: Get guild statistics
      description: Returns statistics for all guilds
      operationId: getGuildAnalytics
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Guild analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildAnalyticsList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /guilds:
    get:
      tags:
        - Guilds
      summary: List guilds connected to the bot
      description: Returns a paginated list of guilds synchronized from Discord along with their availability status.
      operationId: listGuilds
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Guild list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /guilds/{guildId}:
    get:
      tags:
        - Guilds
      summary: Get guild information
      description: Returns information about a specific guild
      operationId: getGuild
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Guild retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guild'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /guilds/{guildId}/settings:
    get:
      tags:
        - Guilds
      summary: Get guild settings
      description: Returns settings for a specific guild
      operationId: getGuildSettings
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildSettings'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Guilds
      summary: Update guild settings
      description: Updates settings for a specific guild
      operationId: updateGuildSettings
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuildSettingsUpdate'
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildSettings'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /guilds/{guildId}/channels:
    get:
      tags:
        - Guilds
      summary: List basic Discord channels
      description: Returns cached Discord channels (text/voice/stage) for a guild so the panel can invoke premium actions.
      operationId: getGuildChannels
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Channel list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuildChannelList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    GuildIdParam:
      name: guildId
      in: path
      description: Discord guild ID
      required: true
      schema:
        type: string
        pattern: '^\d{17,19}$'
    PageParam:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequestError:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
...
    GuildAnalyticsList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/GuildAnalytics'
        pagination:
          $ref: '#/components/schemas/Pagination'
    GuildAnalytics:
      type: object
      properties:
        guildId:
          type: string
        tracksPlayed:
          type: integer
        commandsExecuted:
          type: integer
        activeUsers:
          type: integer
        averageSessionDuration:
          type: integer
        tier:
          type: string
          enum: [FREE, TRIAL, PREMIUM]
    Guild:
      type: object
      properties:
        id:
          type: string
          description: Discord guild ID
        name:
          type: string
          description: Friendly guild name pulled from Discord
        icon:
          type: string
          format: uri
          nullable: true
          description: CDN URL for the guild icon, if available
        memberCount:
          type: integer
          nullable: true
        available:
          type: boolean
          description: Indicates whether the guild is currently reachable by the bot
        subscriptionTier:
          type: string
          description: Plan tier configured in ServerConfiguration
        isPremium:
          type: boolean
          description: True when the guild can access premium-only controls
    GuildList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Guild'
        pagination:
          $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
    GuildChannelList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/GuildChannel'
        timestamp:
          type: string
          format: date-time
    GuildChannel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [text, voice, announcement, stage]
...
    GuildSettings:
      type: object
      properties:
        guildId:
          type: string
        volume:
          type: integer
          minimum: 0
          maximum: 200
        autoplay:
          type: boolean
        autoplayMode:
          type: string
          enum: [similar, artist, genre, mixed]
        loopMode:
          type: string
          enum: [off, track, queue]
        streamable:
          type: boolean
          description: Indicates whether the API can proxy the audio for web playback
        djRoleId:
          type: string
          nullable: true
        notificationChannelId:
          type: string
          nullable: true
    GuildSettingsUpdate:
      type: object
      properties:
        volume:
          type: integer
          minimum: 0
          maximum: 200
        autoplay:
          type: boolean
        autoplayMode:
          type: string
          enum: [similar, artist, genre, mixed]
        loopMode:
          type: string
          enum: [off, track, queue]
        djRoleId:
          type: string
          nullable: true
        notificationChannelId:
          type: string
          nullable: true
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean